requirements.txt
flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3
